- add logging for CreatePipelineJob - log CreateScheduleRequest - log for UpdateScheduel request failure - CreateScheduleRequest success logging

+++ provider-service/vai/internal/provider/job_builder.go
@@ -128,7 +128,7 @@ func (b DefaultJobBuilder) MkSchedule(
 			},
 		},
 		DisplayName:           fmt.Sprintf("rc-%s-%s", rsd.Name.Namespace, rsd.Name.Name),
-		MaxConcurrentRunCount: int64(maxConcurrentRunCount),
+		MaxConcurrentRunCount: maxConcurrentRunCount,
 		AllowQueueing:         true,
 	}
